feat: variable length address id

+++ contracts/AddressRegistry.sol
@@ -2,11 +2,11 @@
 pragma solidity >=0.7.0;
 
 contract AddressRegistry {
-  mapping (uint24 => address) public addressById;
-  mapping (address => uint24) public idByAddress;
-  uint24 public latestId = 0;
+  mapping (uint40 => address) public addressById;
+  mapping (address => uint40) public idByAddress;
+  uint40 public latestId = 0;
 
-  function bindAddress(address a) public returns (uint24 id) {
+  function bindAddress(address a) public returns (uint40 id) {
     // find the number of zeroes in the address
     uint8 zeroCount = 0;
     for (uint8 x = 0; x < 20; x++) {
@@ -18,7 +18,7 @@ contract AddressRegistry {
     require(zeroCount < 16, 'nosav');
     // otherwise we'll store it
     require(idByAddress[a] == 0, 'dupe');
-    require(latestId < type(uint24).max, 'full');
+    require(latestId < type(uint40).max, 'full');
     id = ++latestId;
     addressById[id] = a;
     idByAddress[a] = id;
